Update calculator with decimal calculations

+++ py101/lesson_2/calculator.py
@@ -8,7 +8,7 @@ def prompt(message):
 
 def invalid_number(number_str):
     try:
-        int(number_str)
+        float(number_str)
     except ValueError:
         return True
 
@@ -42,13 +42,13 @@ while calculate:
 
     match operation:
         case '1':   # '1' represents addition
-            output = int(number1) + int(number2)
+            output = float(number1) + float(number2)
         case '2':   #'2' represents subtraction
-            output = int(number1) - int(number2)
+            output = float(number1) - float(number2)
         case '3':   #'3' represents multiplication
-            output = int(number1) * int(number2)
+            output = float(number1) * float(number2)
         case '4':   #'4' represents division
-            output = int(number1) / int(number2)
+            output = float(number1) / float(number2)
 
     prompt(f'The result is {output}')
     prompt(MESSAGE['recalculate'])
